fix: rename httpreceiver to httplogreceiver
