+++ components/components.go
@@ -119,7 +119,7 @@ import (
 	_ "github.com/SigNoz/signoz-otel-collector/pkg/parser/grok"
 	"github.com/SigNoz/signoz-otel-collector/processor/signozspanmetricsprocessor"
 	"github.com/SigNoz/signoz-otel-collector/processor/signoztailsampler"
-	"github.com/SigNoz/signoz-otel-collector/receiver/httpreceiver"
+	"github.com/SigNoz/signoz-otel-collector/receiver/httplogreceiver"
 )
 
 func Components() (otelcol.Factories, error) {
@@ -189,7 +189,7 @@ func Components() (otelcol.Factories, error) {
 		udplogreceiver.NewFactory(),
 		zipkinreceiver.NewFactory(),
 		zookeeperreceiver.NewFactory(),
-		httpreceiver.NewFactory(),
+		httplogreceiver.NewFactory(),
 	}
 	for _, rcv := range factories.Receivers {
 		receivers = append(receivers, rcv)
