+++ receiver/httplogreceiver/bodyparser/logs.go
